+++ alfred-telemetry-platform/src/main/java/eu/xenit/alfred/telemetry/alfrescointegration/AlfredTelemetryMetricsController.java
@@ -3,13 +3,20 @@ package eu.xenit.alfred.telemetry.alfrescointegration;
 import io.micrometer.core.instrument.MeterRegistry;
 import io.micrometer.core.instrument.composite.CompositeMeterRegistry;
 import org.alfresco.enterprise.metrics.MetricsController;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
  * Implementation of Alfresco's {@link MetricsController} that will offer the global MicroMeter registry as registry.
  */
 public class AlfredTelemetryMetricsController implements MetricsController {
 
-    private CompositeMeterRegistry globalRegistry;
+    private static final Logger log = LoggerFactory.getLogger(AlfredTelemetryMetricsController.class);
+
+    private static final String MSG_INTEGRATION_FAILED =
+            "Failed to integrate Alfred Telemetry with the out of the box Alfresco metrics: ";
+
+    private final CompositeMeterRegistry globalRegistry;
     private final MetricsController alfrescoMetricsController;
 
     AlfredTelemetryMetricsController(CompositeMeterRegistry globalRegistry,
@@ -23,7 +30,17 @@ public class AlfredTelemetryMetricsController implements MetricsController {
 
     private void registerAlfrescoControllerInGlobalRegistry(final boolean enableDefaultAlfrescoRegistry) {
         if (enableDefaultAlfrescoRegistry) {
-            globalRegistry.add(alfrescoMetricsController.getRegistry());
+            if (!alfrescoMetricsController.isEnabled()) {
+                log.warn(MSG_INTEGRATION_FAILED
+                        + "out of the box Alfresco metrics not enabled, see ${metrics.enabled} property.");
+                return;
+            }
+            final MeterRegistry alfrescoRegistry = alfrescoMetricsController.getRegistry();
+            if (alfrescoRegistry == null) {
+                log.warn(MSG_INTEGRATION_FAILED + "out of the box Alfresco MeterRegistry not correctly initialized.");
+                return;
+            }
+            globalRegistry.add(alfrescoRegistry);
         }
     }